backend/requirements.txt
# Web Framework (API)
fastapi
uvicorn
python-multipart   # Required for File Uploads (Excel/CSV)
requests           # Needed for Perplexity API calls

# AI & Agents
langgraph
langchain-openai   # For Perplexity API via OpenAI-compatible interface
langchain-core
pydantic           # Data validation and models

# Vector DB & Embeddings
chromadb           # Local vector database
sentence-transformers # Text embeddings (multilingual-e5-small)
pypdf              # PDF file processing

# Observability (Optional - only if LANGSMITH_API_KEY is set)
langsmith          # LangSmith tracing (optional, non-intrusive)

# Data Processing
pandas
python-dotenv

# OCR & Document Processing
pytesseract        # Tesseract OCR wrapper
Pillow             # Image processing

# Testing
pytest             # Test framework
pytest-asyncio     # Async test support
httpx              # HTTP client for API testing

# Database
# sqlite3 is built-in